app/requirements/base.txt
aldryn-apphooks-config==0.3.3
aldryn-boilerplates==0.7.5
aldryn-categories==1.1.0
aldryn-common==1.0.4
aldryn-newsblog==1.3.3
aldryn-people==1.2.2
aldryn-reversion==1.1.0
aldryn-translation-tools==0.2.1
Babel==2.5.1
backport-collections==0.1
cmsplugin-filer==1.1.3
decorator==4.1.2
Django==1.9.13
django-appconf==1.0.2
django-appdata==0.1.6
django-classy-tags==0.8.0
django-cms==3.4.5
django-filer==1.3.0
django-formtools==2.1
django-mptt==0.8.7
django-parler==1.9.1
django-phonenumber-field==1.3.0
django-polymorphic==1.3
django-reversion==1.10.2
django-sekizai==0.10.0
django-sortedm2m==1.5.0
django-taggit==0.22.2
django-treebeard==4.2.0
djangocms-admin-style==1.2.7
djangocms-attributes-field==0.3.0
djangocms-column==1.7.0
djangocms-googlemap==1.1.1
djangocms-link==2.1.2
djangocms-snippet==1.9.2
djangocms-style==2.0.2
djangocms-text-ckeditor==3.5.1
djangocms-video==2.0.4
easy-thumbnails==2.4.2
elasticsearch==6.1.1
gunicorn==19.7.1
html5lib==0.9999999
lxml==4.1.1
pexpect==4.3.1
phonenumbers==8.8.8
phonenumberslite==8.8.8
pickleshare==0.7.4
prompt-toolkit==1.0.15
psycopg2==2.7.3.2
ptyprocess==0.5.2
Pygments==2.2.0
python-dateutil==2.6.1
python-slugify==1.2.0
pytz==2017.3
simplegeneric==0.8.1
six==1.11.0
text-unidecode==1.1
traitlets==4.3.2
Unidecode==0.4.21
urllib3==1.22
wcwidth==0.1.7
YURL==0.13
requests